+++ PlayniteIntegration/AppearanceConverters.cs
@@ -131,4 +131,18 @@ namespace ControllerSessionManager.PlayniteIntegration
             throw new NotSupportedException();
         }
     }
+
+    public sealed class InverseBooleanToVisibilityConverter : IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            var flag = value is bool && (bool)value;
+            return flag ? Visibility.Collapsed : Visibility.Visible;
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            throw new NotSupportedException();
+        }
+    }
 }